+++ devices/src/legacy/serial.rs
@@ -67,10 +67,16 @@ pub struct Serial {
     baud_divisor: u16,
     in_buffer: VecDeque<u8>,
     out: Option<Box<io::Write + Send>>,
+    data_len: usize,
 }
 
 impl Serial {
-    fn new(interrupt_evt: EventFd, out: Option<Box<io::Write + Send>>) -> Serial {
+    fn new(
+        interrupt_evt: EventFd,
+        out: Option<Box<io::Write + Send>>,
+        len: Option<usize>,
+    ) -> Serial {
+        let data_len = len.unwrap_or(1);
         Serial {
             interrupt_enable: 0,
             interrupt_identification: DEFAULT_INTERRUPT_IDENTIFICATION,
@@ -83,17 +89,22 @@ impl Serial {
             baud_divisor: DEFAULT_BAUD_DIVISOR,
             in_buffer: VecDeque::new(),
             out,
+            data_len,
         }
     }
 
     /// Constructs a Serial port ready for output.
-    pub fn new_out(interrupt_evt: EventFd, out: Box<io::Write + Send>) -> Serial {
-        Self::new(interrupt_evt, Some(out))
+    pub fn new_out(
+        interrupt_evt: EventFd,
+        out: Box<io::Write + Send>,
+        data_len: Option<usize>,
+    ) -> Serial {
+        Self::new(interrupt_evt, Some(out), data_len)
     }
 
     /// Constructs a Serial port with no connected output.
-    pub fn new_sink(interrupt_evt: EventFd) -> Serial {
-        Self::new(interrupt_evt, None)
+    pub fn new_sink(interrupt_evt: EventFd, data_len: Option<usize>) -> Serial {
+        Self::new(interrupt_evt, None, data_len)
     }
 
     /// Queues raw bytes for the guest to read and signals the interrupt if the line status would
@@ -195,12 +206,12 @@ impl Serial {
 
 impl BusDevice for Serial {
     fn read(&mut self, offset: u64, data: &mut [u8]) {
-        if data.len() != 1 {
+        if data.len() != self.data_len {
             METRICS.uart.missed_read_count.inc();
             return;
         }
 
-        data[0] = match offset as u8 {
+        data[0] = match (offset / self.data_len as u64) as u8 {
             DLAB_LOW if self.is_dlab_set() => self.baud_divisor as u8,
             DLAB_HIGH if self.is_dlab_set() => (self.baud_divisor >> 8) as u8,
             DATA => {
@@ -227,11 +238,11 @@ impl BusDevice for Serial {
     }
 
     fn write(&mut self, offset: u64, data: &[u8]) {
-        if data.len() != 1 {
+        if data.len() != self.data_len {
             METRICS.uart.missed_write_count.inc();
             return;
         }
-
+        let offset = offset / self.data_len as u64;
         if let Err(e) = self.handle_write(offset as u8, data[0]) {
             error!("Failed the write to serial: {:?}", e);
             METRICS.uart.error_count.inc();
@@ -272,7 +283,7 @@ mod tests {
         let intr_evt = EventFd::new().unwrap();
         let serial_out = SharedBuffer::new();
 
-        let mut serial = Serial::new_out(intr_evt, Box::new(serial_out.clone()));
+        let mut serial = Serial::new_out(intr_evt, Box::new(serial_out.clone()), None);
 
         serial.write(u64::from(DATA), &[b'x', b'y']);
         serial.write(u64::from(DATA), &[b'a']);
@@ -289,8 +300,11 @@ mod tests {
         let intr_evt = EventFd::new().unwrap();
         let serial_out = SharedBuffer::new();
 
-        let mut serial =
-            Serial::new_out(intr_evt.try_clone().unwrap(), Box::new(serial_out.clone()));
+        let mut serial = Serial::new_out(
+            intr_evt.try_clone().unwrap(),
+            Box::new(serial_out.clone()),
+            None,
+        );
 
         // write 1 to the interrupt event fd, so that read doesn't block in case the event fd
         // counter doesn't change (for 0 it blocks)
@@ -323,7 +337,7 @@ mod tests {
     #[test]
     fn serial_thr() {
         let intr_evt = EventFd::new().unwrap();
-        let mut serial = Serial::new_sink(intr_evt.try_clone().unwrap());
+        let mut serial = Serial::new_sink(intr_evt.try_clone().unwrap(), None);
 
         // write 1 to the interrupt event fd, so that read doesn't block in case the event fd
         // counter doesn't change (for 0 it blocks)
@@ -341,7 +355,7 @@ mod tests {
 
     #[test]
     fn serial_dlab() {
-        let mut serial = Serial::new_sink(EventFd::new().unwrap());
+        let mut serial = Serial::new_sink(EventFd::new().unwrap(), None);
 
         serial.write(u64::from(LCR), &[LCR_DLAB_BIT as u8]);
         serial.write(u64::from(DLAB_LOW), &[0x12 as u8]);
@@ -358,7 +372,7 @@ mod tests {
 
     #[test]
     fn serial_modem() {
-        let mut serial = Serial::new_sink(EventFd::new().unwrap());
+        let mut serial = Serial::new_sink(EventFd::new().unwrap(), None);
 
         serial.write(u64::from(MCR), &[MCR_LOOP_BIT as u8]);
         serial.write(u64::from(DATA), &[b'a']);
@@ -380,12 +394,31 @@ mod tests {
 
     #[test]
     fn serial_scratch() {
-        let mut serial = Serial::new_sink(EventFd::new().unwrap());
+        let mut serial = Serial::new_sink(EventFd::new().unwrap(), None);
 
         serial.write(u64::from(SCR), &[0x12 as u8]);
 
         let mut data = [0u8];
         serial.read(u64::from(SCR), &mut data[..]);
         assert_eq!(data[0], 0x12 as u8);
+    }
+
+    #[test]
+    fn test_serial_data_len() {
+        const LEN: usize = 4;
+        let mut serial = Serial::new_sink(EventFd::new().unwrap(), Some(LEN));
+
+        let missed_writes_before = METRICS.uart.missed_write_count.count();
+        // Trying to write data of length different than the one that we initialized the device with
+        // should increase the `missed_write_count` metric.
+        serial.write(u64::from(DATA), &[b'x']);
+        let missed_writes_after = METRICS.uart.missed_write_count.count();
+        assert_eq!(missed_writes_before, missed_writes_after - 1);
+
+        let data = [b'x'; LEN];
+        serial.write(u64::from(DATA), &data);
+        // When we write data that has the length used to initialize the device, the `missed_write_count`
+        // metric stays the same.
+        assert_eq!(missed_writes_before, missed_writes_after - 1);
     }
 }